frasa_db/requirements.txt
PyMySQL==1.0.2
Flask==2.0.1
SQLAlchemy==1.3.20
gunicorn==20.0.4
python-dotenv==0.15.0